+++ src/mkvlib.h
@@ -6,6 +6,8 @@
 #ifndef _JOHN_MKVLIB_H
 #define _JOHN_MKVLIB_H
 
+#include <stdint.h>
+
 #define UNK_STR 255
 
 struct s_pwd {
@@ -17,7 +19,7 @@ struct s_pwd {
 
 extern unsigned char *proba1;
 extern unsigned char *proba2;
-extern unsigned long long *nbparts;
+extern uint64_t *nbparts;
 extern unsigned char *first;
 extern unsigned char charsorted[256 * 256];
 
@@ -25,13 +27,13 @@ extern unsigned int gmax_level;
 extern unsigned int gmax_len;
 extern unsigned int gmin_level;
 extern unsigned int gmin_len;
-extern unsigned long long gidx;
-extern unsigned long long gstart;
-extern unsigned long long gend;
+extern uint64_t gidx;
+extern uint64_t gstart;
+extern uint64_t gend;
 
-void print_pwd(unsigned long long index, struct s_pwd *pwd,
+void print_pwd(uint64_t index, struct s_pwd *pwd,
                unsigned int max_lvl, unsigned int max_len);
-unsigned long long nb_parts(unsigned char lettre, unsigned int len,
+uint64_t nb_parts(unsigned char lettre, unsigned int len,
                             unsigned int level, unsigned int max_lvl, unsigned int max_len);
 void init_probatables(char *filename);
 #endif
