+++ src/rar2john.c
@@ -352,8 +352,8 @@ next_file_header:
 			fprintf(stderr,
 			        "! HEAD_SIZE: %d, PACK_SIZE: %"PRIu64", UNP_SIZE: %"PRIu64"\n",
 			        file_header_head_size,
-			        (unsigned long long)file_header_pack_size,
-			        (unsigned long long)file_header_unp_size);
+			        (uint64_t)file_header_pack_size,
+			        (uint64_t)file_header_unp_size);
 			fprintf(stderr, "! file_header_block:\n!  ");
 			for (i = 0; i < 32; ++i)
 				fprintf(stderr, " %02x", file_header_block[i]);
@@ -554,8 +554,8 @@ next_file_header:
 		}
 
 		best_len += sprintf(&best[best_len], "*%"PRIu64"*%"PRIu64"*",
-		        (unsigned long long)file_header_pack_size,
-		        (unsigned long long)file_header_unp_size);
+		        (uint64_t)file_header_pack_size,
+		        (uint64_t)file_header_unp_size);
 
 		/* We always store it inline */
